+++ configure.py
@@ -630,7 +630,6 @@ scylla_tests = set([
     'test/unit/lsa_sync_eviction_test',
     'test/unit/row_cache_alloc_stress_test',
     'test/unit/row_cache_stress_test',
-    'test/unit/bptree_compaction_test',
     'test/unit/cross_shard_barrier_test',
 ])
 
@@ -1394,7 +1393,6 @@ pure_boost_tests = set([
     'test/boost/small_vector_test',
     'test/boost/top_k_test',
     'test/boost/vint_serialization_test',
-    'test/boost/bptree_test',
     'test/boost/utf8_test',
     'test/boost/string_format_test',
     'test/manual/streaming_histogram_test',
@@ -1415,7 +1413,6 @@ tests_not_using_seastar_test_framework = set([
     'test/unit/lsa_async_eviction_test',
     'test/unit/lsa_sync_eviction_test',
     'test/unit/row_cache_alloc_stress_test',
-    'test/unit/bptree_compaction_test',
     'test/manual/sstable_scan_footprint_test',
     'test/unit/cross_shard_barrier_test',
 ]) | pure_boost_tests
